+++ packages/data/src/entry-form/default-entry-output.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Field } from '@ec.components/core/src/field/field';
+import { Item } from '@ec.components/core';
 
 /** This component holds the templates for all output types that cannot be represented by the default output component. */
 @Component({
@@ -9,4 +10,6 @@ import { Field } from '@ec.components/core/src/field/field';
 export class DefaultEntryOutputComponent {
   /** The field to output. */
   field: Field;
+  /** The item to display */
+  item: Item<any>;
 }