+++ packages/data/src/files/image/image.directive.ts
@@ -1,9 +1,9 @@
-/**
- * Created by felix on 23.05.17.
- */
 import { Directive, ElementRef, Input, OnChanges } from '@angular/core';
 import { AssetDirective } from '../asset/asset.directive';
 import { SdkService } from '../../sdk/sdk.service';
+import DMAssetResource from 'ec.sdk/lib/resources/publicAPI/DMAssetResource';
+import PublicAssetResource from 'ec.sdk/lib/resources/publicAPI/PublicAssetResource';
+import PublicAPI from 'ec.sdk/lib/PublicAPI';
 
 /** Loads an public asset image by id to the template. It can be used with img's to auto load the url to the src. */
 @Directive({
@@ -16,19 +16,20 @@ export class ImageDirective extends AssetDirective implements OnChanges {
   /** The size that should be requested. */
   @Input() size = 200;
   /** The assetID that should be loaded */
-  @Input() ecImage: string;
+  @Input() ecImage: string | DMAssetResource;
   /** Resolved asset url. */
   url: string;
 
   constructor(public sdk: SdkService, private elementRef: ElementRef) {
     super(sdk);
-
   }
 
   /** Reads ecImage input as assetId and loads if autoload is not false */
   ngOnChanges() {
-    if (this.ecImage) {
+    if (typeof this.ecImage === 'string') {
       this.assetId = this.ecImage;
+    } else if (this.ecImage && this.ecImage.assetID) {
+      this.use(this.ecImage);
     }
     if (this.autoload !== false && this.assetId) {
       this.load();
@@ -37,18 +38,29 @@ export class ImageDirective extends AssetDirective implements OnChanges {
 
   /** Calls super.load, then resolves the image url and assigns it to the native element src (only if it is an img) */
   load(id?: string) {
-    return super.load(id).then((asset) => {
+    return super.load(id).then(this.use.bind(this));
+  }
+
+  use(asset: PublicAssetResource | DMAssetResource) {
+    return Promise.resolve().then(() => {
       if (asset.type !== 'image') {
         return Promise.reject(`ecImage only works for assets of type image.
         Loaded id ${asset.id} is of type ${asset.type}`);
       }
-      if (this.thumb) {
-        return asset.getImageThumbUrl(this.size);
+      if (asset instanceof DMAssetResource) { // new asset
+        return asset.getFileVariant(this.size, this.thumb);
+      } else if (asset instanceof PublicAssetResource) { // old asset
+        if (this.thumb) {
+          return asset.getImageThumbUrl(this.size, '');
+        }
+        return asset.getImageUrl(this.size, '');
       }
-      return asset.getImageUrl(this.size)
-    }).then((url) => {
-      this.url = url;
-      this.elementRef.nativeElement.src = this.url;
-    });
+    }).then(this.setUrl.bind(this));
   }
+
+  setUrl(url: string) {
+    this.url = url;
+    this.elementRef.nativeElement.src = this.url;
+  }
+
 }